refactor tracking specific code from layout.svelte into analytics.svelte
